pyproject.toml
[project]
name = "rc-car-controls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "opencv-python>=4.12.0.88",
    "psutil>=7.0.0",
    "pyserial>=3.5",
]
